Fixed searchFood close to edge
Related to #6
If a Dot was close to the edge, searchFood would calculate the interval in food-vector correctly. But this meant that the end of the interval was somewhere in the beginning of the vector and the start somewhere at the end: |-.--->         <--|
Now this case gets caught and the minimum is calculated from 0 to end and start to the end of the vector.