[Not Compiling] searchFood() with grid
Tried to implement the answer from https://stackoverflow.com/a/59432406/12540220
I have some problems with sizes and indexes of vectors. This means that the program just crashes without any error message at the moment. I've got a `std::bad_alloc` once, which confirms that the error is somewhere where I mentioned.
As the commit from the cannibalism-branch aren't merged yet, I also set the get-Functions to const here. Hopefully this wont lead to conflicts in the future...